refactor: replace img tag with Next.js Image and move Hero components to Hero folder
