+++ src/app/(frontend)/_components/sections/HomePage/Hero/CTAButton.tsx
@@ -1,20 +1,17 @@
-'use client';
+'use client'
 export function CTAButton() {
   return (
     <a
       href="https://form.jotform.com/250418674375867?fbclid=PAZXh0bgNhZW0CMTEAAaeWIjTTV9xmRZdfLddy8HFmM9hUlfwNq9s9cwQ25cArwsCTzYgQgbH-2bx3Pw_aem_0HuEKOXK5sj-2w6iUQDzWA"
       target="_blank"
       rel="noopener noreferrer"
-      className="inline-block mt-8"
+      className="mt-8 inline-block"
     >
-      <button
-        className="px-8 py-3 bg-white/90 text-black rounded-3xl shadow-lg hover:bg-orange-400 hover:text-white transition"
-      >
+      <button className="rounded-3xl bg-white/90 px-8 py-3 text-black shadow-lg transition hover:bg-orange-400 hover:text-white">
         Become a member
       </button>
     </a>
-  );
+  )
 }
-
 
 // This button is used in the HeroContent component to navigate to the membership page.
