+++ src/app/(frontend)/_components/sections/HomePage/Hero/HeroSection.tsx
@@ -1,15 +1,18 @@
-"use client";
+'use client'
 
-import { BackgroundImage } from '../../../ui/BackgroundImage';
-import { Overlay } from '../../../ui/Overlay';
-import { HeroContent } from './HeroContent';
+import { BackgroundImage } from './BackgroundImage'
+import { HeroContent } from './HeroContent'
+import { Overlay } from './Overlay'
 
 export default function HeroSection() {
   return (
-    <section className="relative min-h-screen flex items-center justify-center bg-gray-900">
-      <BackgroundImage src="/hero_background_Image.jpg" alt="Kayaking in New Zealand" />
+    <section className="relative flex min-h-screen items-center justify-center bg-gray-900">
+      <BackgroundImage
+        src="/hero_background_Image.jpg"
+        alt="Kayaking in New Zealand"
+      />
       <Overlay />
       <HeroContent />
     </section>
-  );
+  )
 }
