+++ src/app/(frontend)/_components/sections/HomePage/Hero/Overlay.tsx
@@ -1,0 +1,3 @@
+export function Overlay() {
+  return <div className="bg-gray/40 absolute inset-0 z-10" />
+}
