--- src/app/(frontend)/_components/ui/Overlay.tsx
@@ -1,3 +1,0 @@
-export function Overlay() {
-  return <div className="absolute inset-0 bg-gray/40 z-10" />;
-}
